+++ R/burndown.R
@@ -1,5 +1,47 @@
 #' Calculate storypoints ans status of the Cards of a specific Board
 #'
+#' Calculates storypoints of released cards on a specific board
+#'
+#' @param organisation A character. The name of the organisation as specified in the URL:
+#' https://<organisation>.upwave.io
+#' @param board_id An integer. The number of the board you want to inspect as specified in
+#' the URL: https://<organisation>.upwave.io/board/<board_id>/view/
+#' @param token A character. The API token to access boards on upwave.io. The token can be
+#' obtained by visiting your account settings. Click your profile image, select "Settings"
+#' and find your API-Key in the "Account" tab.
+#'
+#' @return A list containing a data.frame with all storypoints and the status per card and
+#' the numeric sum_storypoints.
+#' @export
+#'
+#' @examples upwaver::storypoints_and_status_released(
+#' "ims-fhs", 14351, "a44fa67c5df2acc9836058ffca870d7b78b017cb")
+storypoints_and_status_released <- function(organisation, board_id, token) {
+  cards <- upwaver::list_cards(organisation, board_id, token)$results
+  if (length(cards) >= 100) {stop("More than 100 cards on board, calculation of storypoints may be wrong.")}
+
+  df <- data.frame(title = character(length(cards)),
+                   status = numeric(length(cards)),
+                   storypoints = character(length(cards)),
+                   stringsAsFactors = F)
+  for (i in 1:length(cards)) {
+    # state 0: Not specified, 1: Not started, 2: In progress, 3: Completed
+    df[i, 1] <- cards[[i]]$title
+    df[i, 2] <- cards[[i]]$state
+    if (cards[[i]]$state == 3) {
+      description <- cards[[i]]$description
+      df[i, 3] <- aufwand(description)
+    } else {
+      df[i, 3] <- "Active" # active tasks
+    }
+  }
+  return(list(storypoints_status = df,
+              sum_storypoints = sum(as.numeric(df$storypoints[df$storypoints != "Active"]))))
+}
+
+
+#' Calculate storypoints ans status of the Cards of a specific Board
+#'
 #' Calculates storypoints (of active Cards, which are Cards that haven't been completed
 #' yet) and status in the backlog of an 'Upwave' board.
 #' Storypoints have to be specified in the description (directly beneath the title) of
@@ -21,6 +63,8 @@
 #' @examples upwaver::storypoints_and_status("ims-fhs", 14351, "a44fa67c5df2acc9836058ffca870d7b78b017cb")
 storypoints_and_status <- function(organisation, board_id, token) {
   cards <- upwaver::list_cards(organisation, board_id, token)$results
+
+  if (length(cards) >= 100) {stop("More than 100 cards on board, calculation of storypoints may be wrong.")}
 
   df <- data.frame(title = character(length(cards)),
                    status = numeric(length(cards)),